Fix migrate for database

+++ db/migrate/20260106122023_remove_content_from_message.rb
@@ -1,5 +1,5 @@
 class RemoveContentFromMessage < ActiveRecord::Migration[8.1]
   def change
-    remove_column :messages, :content, :string
+    remove_column :messages, :content, :string if column_exists?(:messages, :content)
   end
 end
